requirements.txt
# 依赖库配置
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
dash>=2.14.0
dash-bootstrap-components>=1.5.0
pyyaml>=6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pymysql>=1.1.0
requests>=2.31.0
openpyxl>=3.1.0
xlrd>=2.0.0
kaleido>=0.2.1
python-dateutil>=2.8.0

# 可选依赖
pillow>=10.0.0
reportlab>=4.0.0
